+++ sdk/internal/CHANGELOG.md
@@ -12,6 +12,8 @@
 ### Bugs Fixed
 
 ### Other Changes
+* Zero `RecordingOptions.ProxyPort` is interpreted as indicating the default port used
+  by `StartTestProxy`
 
 ## 1.10.0 (2024-07-16)